Create Solution.py
The local newspaper, Stavanger Aftenblad, ran an article on September 4th about one of
the long-time top managers in Equinor, Margareth Øvrum, who is about to retire at age 62. When she
was 32, Øvrum became the first female platform manager in Statoil. Since then she has held a number
of top management jobs in the company and since 2004 she has been a member of the company’s
executive management team reporting directly to the managing director. Her current job is as
managing director of Equinor Brazil.
According to the newspaper, the present value at the end of 2019 of Øvrum’s lifetime retirement
pension was NOK 69 million. According to NAV1, life expectancy for a woman born in 1958 is about 88
years.
The Norwegian State’s Pension Fund use an annual discount rate of 4%2 to calculate the value of the
pension fund (assume annual compounding).
a) Given the information above, what will be the size of Øvrum’s monthly pension when she
begins her retirement in January 2020?
b) Use Python to do a sensitivity analysis to evaluate what her monthly pension will be for
discount rates between 0 and 5%. Graph your results.